+++ controllers/sms/index.js
@@ -1,0 +1,7 @@
+const { createSmsValidation } = require('./createSmsValidation');
+const { createCertificateBySmsCode } = require('./createCertificateBySmsCode');
+
+module.exports = {
+  createSmsValidation,
+  createCertificateBySmsCode,
+};
