+++ routes/SmsRoutes.js
@@ -1,18 +1,8 @@
 /* eslint-disable max-len */
-/* eslint-disable no-await-in-loop */
-/* eslint-disable no-restricted-syntax */
 const router = require('express').Router();
-const ResponseHandler = require('../utils/ResponseHandler');
-const Certificate = require('../models/Certificate');
 
-const SmsService = require('../services/SmsService');
-const UserService = require('../services/UserService');
-const MouroService = require('../services/MouroService');
-const CertService = require('../services/CertService');
-
+const sms = require('../controllers/sms');
 const { validateBody, checkValidationResult } = require('../utils/Validator');
-const CodeGenerator = require('../utils/CodeGenerator');
-const Messages = require('../constants/Messages');
 const Constants = require('../constants/Constants');
 const { halfHourLimiter } = require('../policies/RateLimit');
 
@@ -68,39 +58,7 @@ router.post(
   ]),
   checkValidationResult,
   halfHourLimiter,
-  async (req, res) => {
-    const phoneNumber = await UserService.normalizePhone(req.body.cellPhoneNumber);
-    const { did } = req.body;
-    const { password } = req.body;
-    const { unique } = req.body;
-
-    try {
-      // Validar que el teléfono no esté en uso
-      if (unique) await UserService.telTaken(phoneNumber, did);
-
-      // Si se ingresó contraseña, validarla
-      if (password && did) await UserService.getAndValidate(did, password);
-
-      // Generar código de validación
-      const code = CodeGenerator.generateCode(Constants.RECOVERY_CODE_LENGTH);
-      // eslint-disable-next-line no-console
-      if (Constants.DEBUGG) console.log(code);
-
-      // Crear y guardar pedido de validación de teléfono
-      await SmsService.create(phoneNumber, code, undefined);
-
-      // Mandar sms con el código de validacion
-      if (Constants.NO_SMS) {
-        return ResponseHandler.sendRes(res, { code });
-      }
-
-      await SmsService.sendValidationCode(phoneNumber, code);
-
-      return ResponseHandler.sendRes(res, Messages.SMS.SUCCESS.SENT);
-    } catch (err) {
-      return ResponseHandler.sendErr(res, err);
-    }
-  },
+  sms.createSmsValidation,
 );
 
 /**
@@ -149,49 +107,7 @@ router.post(
   ]),
   checkValidationResult,
   halfHourLimiter,
-  async (req, res) => {
-    const cellPhoneNumber = await UserService.normalizePhone(req.body.cellPhoneNumber);
-
-    const { validationCode } = req.body;
-    const { did } = req.body;
-
-    try {
-      // Validar código
-      const phone = await SmsService.isValid(cellPhoneNumber, validationCode);
-
-      // Validar que no existe un usuario con ese mail
-      const user = await UserService.getByTel(cellPhoneNumber);
-      if (user) return ResponseHandler.sendErr(res, Messages.SMS.ERR.ALREADY_EXISTS);
-
-      // Generar certificado validando que ese did le corresponde al dueño del teléfono
-      const cert = await CertService.createPhoneCertificate(did, cellPhoneNumber);
-      await CertService.verifyCertificatePhoneNumber(cert);
-
-      // Revocar certificado anterior
-      const old = await Certificate.findByType(did, Constants.CERTIFICATE_NAMES.TEL);
-      for (const elem of old) {
-        elem.update(Constants.CERTIFICATE_STATUS.REVOKED);
-        const jwt = await elem.getJwt();
-        await MouroService.revokeCertificate(jwt, elem.hash, did);
-      }
-
-      // Mandar certificado a mouro
-      const jwt = await MouroService.saveCertificate(cert, did);
-
-      // Validar código y actualizar pedido de validación de teléfono
-      await Certificate.generate(
-        Constants.CERTIFICATE_NAMES.TEL,
-        did,
-        Constants.CERTIFICATE_STATUS.UNVERIFIED,
-        jwt.data,
-        jwt.hash,
-      );
-      await SmsService.validatePhone(phone, did);
-      return ResponseHandler.sendRes(res, Messages.SMS.SUCCESS.MATCHED(cert));
-    } catch (err) {
-      return ResponseHandler.sendErr(res, err);
-    }
-  },
+  sms.createCertificateBySmsCode,
 );
 
 module.exports = router;
